Solving problems when compiling in Windows

+++ applications/DEMApplication/custom_strategies/strategies/explicit_solver_continuum.cpp
@@ -648,14 +648,14 @@ namespace Kratos {
 
         KRATOS_TRY
 
-        const unsigned int maximum_allowed_number_of_intact_bonds = (unsigned int) GetModelPart().GetProcessInfo()[MAX_NUMBER_OF_INTACT_BONDS_TO_CONSIDER_A_SPHERE_BROKEN];
+        const int maximum_allowed_number_of_intact_bonds = (int) GetModelPart().GetProcessInfo()[MAX_NUMBER_OF_INTACT_BONDS_TO_CONSIDER_A_SPHERE_BROKEN];
 
         #pragma omp parallel for
-        for (unsigned int i = 0; i < mListOfSphericContinuumParticles.size(); i++) {
+        for (int i = 0; i < (int) mListOfSphericContinuumParticles.size(); i++) {
 
-            unsigned int number_of_intact_bonds = 0;
+            int number_of_intact_bonds = 0;
 
-            for (unsigned int j = 0; j < mListOfSphericContinuumParticles[i]->mContinuumInitialNeighborsSize; j++) {
+            for (int j = 0; j < (int) mListOfSphericContinuumParticles[i]->mContinuumInitialNeighborsSize; j++) {
 
                 if (!mListOfSphericContinuumParticles[i]->mIniNeighbourFailureId[j]) number_of_intact_bonds++;
                 if (number_of_intact_bonds > maximum_allowed_number_of_intact_bonds) break;
@@ -663,7 +663,7 @@ namespace Kratos {
 
             if (number_of_intact_bonds <= maximum_allowed_number_of_intact_bonds) {
 
-                for (unsigned int j = 0; j < mListOfSphericContinuumParticles[i]->mContinuumInitialNeighborsSize; j++) {
+                for (int j = 0; j < (int) mListOfSphericContinuumParticles[i]->mContinuumInitialNeighborsSize; j++) {
 
                     if (!mListOfSphericContinuumParticles[i]->mIniNeighbourFailureId[j]) mListOfSphericContinuumParticles[i]->mIniNeighbourFailureId[j] = 8;
                 }
